Further cleanups
